AI_project4
Implementation of the code for learning an artificial neural network by means of a genetic algorithm. 
"Neural networks are traditionally learned with the backpropagation algorithm. However, learning a neural network is an optimization problem which can also be tackled by a genetic algorithm. The core idea of approach will be to define a population of neural networks (where each chromosome will be an instsance of a neural network), and then to optimize the network weight values with respect to the error of a neural network on the training set. Once optimized, the learned network should be evaluated on a thus far unseen test set."